refactor(web): to new generic pixi container
