+++ apps/web/src/app/canvases/_utils/use-disable-wheel-zoom.hook.ts
@@ -1,0 +1,17 @@
+import { useCallback } from 'react';
+
+export const useDisableWheelZoom = () => {
+  const disableZoomCallback = useCallback((element: HTMLElement) => {
+    if (element) {
+      element.addEventListener(
+        'wheel',
+        (e) => {
+          e.preventDefault();
+        },
+        { passive: false }
+      );
+    }
+  }, []);
+
+  return { disableZoomCallback };
+};
