+++ apps/web/src/app/canvases/solar-system/solar-system.container.tsx
@@ -1,21 +1,14 @@
-import { Box, Theme, useTheme } from '@chakra-ui/react';
-import { hexStringToNumber } from '@idleverse/theme';
-import { Stage } from '@inlet/react-pixi';
 import { useEffect, useState } from 'react';
 
 import { loadPlanets } from '../../asset-loading/load-planets';
 import { Loading } from '../../components/loading';
-import { animateVar } from '../../_state/reactive-variables';
-import { useResize } from '../_utils/use-resize.hook';
+import { PixiWrapper } from '../_utils/pixi-wrapper';
+
 import { SolarSystem } from './solar-system';
 import { SolarSystemControls } from './ui/controls';
 
 export const SolarSystemContainer = () => {
-  const size = useResize();
-
   const [celestialSpritesLoading, setCelestialSpritesLoading] = useState(true);
-
-  const { colors } = useTheme<Theme>();
 
   useEffect(() => {
     loadPlanets().then(() => setCelestialSpritesLoading(false));
@@ -27,18 +20,8 @@ export const SolarSystemContainer = () => {
     );
 
   return (
-    <Box position="relative">
-      <Stage
-        {...size}
-        options={{
-          backgroundColor: hexStringToNumber(colors.gray['800']),
-          antialias: true,
-        }}
-        onUnmount={() => animateVar(false)}
-      >
-        <SolarSystem />
-      </Stage>
-      <SolarSystemControls />
-    </Box>
+    <PixiWrapper ui={<SolarSystemControls />}>
+      <SolarSystem />
+    </PixiWrapper>
   );
 };
