+++ apps/web/src/app/canvases/solar-system/solar-system.tsx
@@ -6,7 +6,7 @@ import { useEffect, useRef } from 'react';
 import { solarSystemConfigVar, timeVar } from '../../_state/reactive-variables';
 import { useFpsTracker } from '../galaxy-generator/utils/fps-counter';
 import { useResize } from '../_utils/use-resize.hook';
-import { useViewport } from '../_utils/use-viewport';
+import { useViewport } from '../_utils/use-viewport.hook';
 
 import {
   createAnimatedPlanetSprite,
